+++ Classes/DependencyInjection/Widgets/ActionsPerMonthRegistration.php
@@ -15,7 +15,6 @@ use Brotkrueml\MatomoWidgets\Extension;
 use Brotkrueml\MatomoWidgets\Widgets\Provider\GenericBarChartDataProvider;
 use Symfony\Component\DependencyInjection\Reference;
 use TYPO3\CMS\Backend\View\BackendViewFactory;
-use TYPO3\CMS\Core\Information\Typo3Version;
 use TYPO3\CMS\Dashboard\Widgets\BarChartWidget;
 
 /**
@@ -74,7 +73,6 @@ final class ActionsPerMonthRegistration extends AbstractRegistration
         $configuration = $this->services
             ->set($this->buildServiceWidgetId(), BarChartWidget::class)
             ->arg('$dataProvider', new Reference($this->buildServiceDataProviderId()))
-            ->arg('$view', new Reference('dashboard.views.widget'))
             ->arg(
                 '$options',
                 [
@@ -95,10 +93,6 @@ final class ActionsPerMonthRegistration extends AbstractRegistration
                 ],
             );
 
-        if ((new Typo3Version())->getMajorVersion() < 12) {
-            $configuration->arg('$view', new Reference('dashboard.views.widget'));
-        } else {
-            $configuration->arg('$backendViewFactory', new Reference(BackendViewFactory::class));
-        }
+        $configuration->arg('$backendViewFactory', new Reference(BackendViewFactory::class));
     }
 }