+++ Classes/DependencyInjection/Widgets/BrowserPluginsRegistration.php
@@ -101,7 +101,6 @@ final class BrowserPluginsRegistration extends AbstractRegistration
         $this->services
             ->set($this->buildServiceWidgetId(), TableWidget::class)
             ->arg('$dataProvider', new Reference($this->buildServiceDataProviderId()))
-            ->arg('$view', new Reference('dashboard.views.widget'))
             ->arg(
                 '$options',
                 [