+++ Classes/DependencyInjection/Widgets/BrowsersRegistration.php
@@ -15,7 +15,6 @@ use Brotkrueml\MatomoWidgets\Extension;
 use Brotkrueml\MatomoWidgets\Widgets\Provider\GenericDoughnutChartDataProvider;
 use Symfony\Component\DependencyInjection\Reference;
 use TYPO3\CMS\Backend\View\BackendViewFactory;
-use TYPO3\CMS\Core\Information\Typo3Version;
 use TYPO3\CMS\Dashboard\Widgets\DoughnutChartWidget;
 
 /**
@@ -77,7 +76,6 @@ final class BrowsersRegistration extends AbstractRegistration
         $configuration = $this->services
             ->set($this->buildServiceWidgetId(), DoughnutChartWidget::class)
             ->arg('$dataProvider', new Reference($this->buildServiceDataProviderId()))
-            ->arg('$view', new Reference('dashboard.views.widget'))
             ->arg(
                 '$options',
                 [
@@ -99,11 +97,7 @@ final class BrowsersRegistration extends AbstractRegistration
                 ],
             );
 
-        if ((new Typo3Version())->getMajorVersion() < 12) {
-            $configuration->arg('$view', new Reference('dashboard.views.widget'));
-        } else {
-            $configuration->arg('$backendViewFactory', new Reference(BackendViewFactory::class));
-        }
+        $configuration->arg('$backendViewFactory', new Reference(BackendViewFactory::class));
     }
 
     private function buildReportLink(): string