+++ Classes/DependencyInjection/Widgets/JavaScriptErrorsRegistration.php
@@ -103,7 +103,6 @@ final class JavaScriptErrorsRegistration extends AbstractRegistration
         $this->services
             ->set($this->buildServiceWidgetId(), JavaScriptErrorsWidget::class)
             ->arg('$dataProvider', new Reference($this->buildServiceDataProviderId()))
-            ->arg('$view', new Reference('dashboard.views.widget'))
             ->arg(
                 '$options',
                 [