+++ Classes/DependencyInjection/Widgets/PagesNotFoundRegistration.php
@@ -115,7 +115,6 @@ final class PagesNotFoundRegistration extends AbstractRegistration
         $this->services
             ->set($this->buildServiceWidgetId(), TableWidget::class)
             ->arg('$dataProvider', new Reference($this->buildServiceDataProviderId()))
-            ->arg('$view', new Reference('dashboard.views.widget'))
             ->arg(
                 '$options',
                 [