+++ Classes/Extension.php
@@ -18,6 +18,8 @@ final class Extension
 {
     public const KEY = 'matomo_widgets';
 
+    public const JS_IMPORT_PREFIX = '@brotkrueml/matomo-widgets';
+
     private const LANGUAGE_PATH = 'LLL:EXT:' . self::KEY . '/Resources/Private/Language/';
     public const LANGUAGE_PATH_DASHBOARD = self::LANGUAGE_PATH . 'Dashboard.xlf';
     public const LANGUAGE_PATH_SITECONF = self::LANGUAGE_PATH . 'SiteConfiguration.xlf';