+++ ecs.php
@@ -2,6 +2,10 @@
 
 declare (strict_types=1);
 
+use PhpCsFixer\Fixer\Comment\HeaderCommentFixer;
+use PhpCsFixer\Fixer\ClassNotation\ClassAttributesSeparationFixer;
+use PhpCsFixer\Fixer\Strict\DeclareStrictTypesFixer;
+use PHP_CodeSniffer\Standards\Generic\Sniffs\CodeAnalysis\AssignmentInConditionSniff;
 use Symplify\EasyCodingStandard\Config\ECSConfig;
 
 return static function (ECSConfig $config): void {
@@ -19,19 +23,19 @@ HEADER;
         __DIR__ . '/Configuration',
         __DIR__ . '/Tests',
     ]);
-    $config->ruleWithConfiguration(\PhpCsFixer\Fixer\Comment\HeaderCommentFixer::class, [
+    $config->ruleWithConfiguration(HeaderCommentFixer::class, [
         'comment_type' => 'comment',
         'header' => $header,
         'separate' => 'both',
     ]);
     $config->skip([
-        \PhpCsFixer\Fixer\ClassNotation\ClassAttributesSeparationFixer::class => [
+        ClassAttributesSeparationFixer::class => [
             __DIR__ . '/Classes/Extension.php',
         ],
-        \PhpCsFixer\Fixer\Strict\DeclareStrictTypesFixer::class => [
+        DeclareStrictTypesFixer::class => [
             __DIR__ . '/Configuration/SiteConfiguration/Overrides/sites.php',
         ],
-        \PHP_CodeSniffer\Standards\Generic\Sniffs\CodeAnalysis\AssignmentInConditionSniff::class . '.FoundInWhileCondition' => [
+        AssignmentInConditionSniff::class . '.FoundInWhileCondition' => [
             __DIR__ . '/Classes/Domain/Repository/BackendUserGroupRepository',
             __DIR__ . '/Classes/Domain/Repository/DashboardRepository.php',
         ],