+++ ext_emconf.php
@@ -6,11 +6,11 @@ $EM_CONF[$_EXTKEY] = [
     'author' => 'Chris Müller',
     'author_email' => 'typo3@krue.ml',
     'state' => 'stable',
-    'version' => '2.5.0-dev',
+    'version' => '3.0.0-dev',
     'constraints' => [
         'depends' => [
             'php' => '8.1.0-0.0.0',
-            'typo3' => '11.5.0-12.4.99',
+            'typo3' => '12.4.0-13.4.99',
             'dashboard' => '',
         ],
         'conflicts' => [],