+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Compatibility with TYPO3 v13
+
+### Removed
+- Compatibility with TYPO3 v11 (#40)
+
 ## [2.4.0] - 2024-01-16
 
 ### Added